Using `asyncio.gather()` (not recommended)
When one uses `gather()`, one can encounter issues with raised exceptions. For example, suppose the task that sleeps for 1 second raises an `Exception`. This exception will result be presented; however, the other tasks **may not be terminated properly**.

The "Tech with Tim" host reports this issue; however, he was not using Python 3.11. When I ran an example using Python 3.11, all tasks seemed to terminate immediately. I'm uncertain if the behavior that I encountered was a "fix" introduced after the video was recorded or my tasks simply did not have enough of a delay to be noticed.

The recommendation: use `asyncio.TaskGroup()` instead.